Sync upstream with downstream

+++ index.js
@@ -1,41 +1,52 @@
 import fs from 'fs';
 import subprocess from 'child_process';
 import net from 'net';
-import signalhandler from './signalhandler.cjs';
+import { interpretmessage } from './signalhandler.js';
 import { parse } from 'jsonc-parser';
-const { interpretmessage } = signalhandler;
 
 let config = parse(fs.readFileSync('config.jsonc', 'utf8'));
 if (config.phonenumber === '' || config.phonenumber === null || config.phonenumber === undefined) {
     console.error('Phone number not found in config.jsonc!');
-    process.exit(1); // skipcq: JS-0263
+    process.exit(1);
 }
-const phonenumber = config.phonenumber;
 if (config.socketpath === '' || config.socketpath === null || config.socketpath === undefined) {
     console.error('Socket path not found in config.jsonc!');
-    process.exit(1); // skipcq: JS-0263
+    process.exit(1);
 };
 const socketpath = config.socketpath;
 const botname = config.botname;
 const botversion = config.botversion;
 const botavatar = config.botavatar;
 const botabout = config.botabout;
+const phonenumber = config.phonenumber;
+const externalsignal = config.externalsignal;
 config = undefined;
-let daemon; // skipcq: JS-0119
+let daemon;
+
+
+function startconn(client, callback) {
+    if (socketpath.includes(':')) {
+        const [host, port] = socketpath.split(':');
+        client.connect(parseInt(port), host, callback);
+    } else {
+        client.connect(socketpath, callback);
+    }
+}
 
 function gracefulShutdown() {
     console.log('Shutting down TritiumBot...');
-    if (daemon) {
+    if (daemon && !externalsignal) {
         daemon.on('exit', () => {
-            process.exit(0); // skipcq: JS-0263
+            process.exit(0);
         });
         daemon.kill('SIGTERM');
         setTimeout(() => {
             console.error('Could not shut down TritiumBot gracefully, forcefully shutting down...');
-            process.exit(1); // skipcq: JS-0263
+            daemon.kill('SIGKILL');
+            process.exit(1);
         }, 5000);
     } else {
-        process.exit(0); // skipcq: JS-0263
+        process.exit(0);
     }
 };
 
@@ -43,26 +54,31 @@ process.on('SIGTERM', gracefulShutdown);
 process.on('SIGINT', gracefulShutdown);
 
 function signalclidaemon() {
-    const daemon = subprocess.exec(`signal-cli daemon --socket ${socketpath} --receive-mode on-connection`, (err, stdout, _stderr) => { // skipcq: JS-0123
-        if (err) {
-            console.error(err);
-            return;
-        }
-        console.log(stdout);
-    });
-    return daemon;
+    if (!externalsignal) {
+        const sf = socketpath.includes(':') ? '--tcp' : '--socket';
+        const daemon = subprocess.exec(`signal-cli --config ./config daemon ${sf} ${socketpath} --receive-mode on-connection`, (err, stdout, _stderr) => {
+            if (err) {
+                console.error(err);
+                return;
+            }
+            console.log(stdout);
+        });
+        return daemon;
+    } else {
+        return null;
+    }
 };
 
 function signalclihook() {
     const client = new net.Socket();
-    client.connect(socketpath, () => {
+    startconn(client, () => {
         console.log('Signal CLI hook connected');
         setupbotprofile();
     });
     client.on('data', (data) => {
         const message = data.toString();
         if (message === '' || message === null || message === undefined || message === '\n') {
-            return; // skipcq: JS-W1045
+            return;
         } else {
             interpretmessage(message);
         }
@@ -75,9 +91,9 @@ function signalclihook() {
     });
 };
 
-function setupbotprofile() {
+export function setupbotprofile() {
     const client = new net.Socket();
-    client.connect(socketpath, () => {
+    startconn(client, () => {
         const tid = Math.floor(Math.random() * 1024) + 1;
         const id = tid.toString();
         let json = {
@@ -85,6 +101,7 @@ function setupbotprofile() {
             id,
             method: 'updateProfile',
             params: {
+                account: phonenumber,
                 givenName: botname,
                 avatar: botavatar,
                 about: botabout,
@@ -92,43 +109,11 @@ function setupbotprofile() {
         }
         if (botversion === true) {
             const botversionjson = parse(fs.readFileSync('package.json', 'utf8')).version;
-            json.params.familyName = `[${botversionjson} Alpha]`;
+            json.params.familyName = `[${botversionjson} Development]`;
         }
         json = JSON.stringify(json);
         client.write(json);
         client.end();
-        /*client.on('data', (data) => {
-            const content = data.toString();
-            if (content == null || content === '' || content === undefined || content === '\n') {
-                return;
-            } else {
-                try {
-                    const parsedJson = JSON.parse(content);
-                    if (parsedJson.id === id) {
-                        if (parsedJson.error) {
-                            console.error('Error sending profile data:', parsedJson.error, parsedJson.error.data.response.results);
-                            return;
-                        }
-                        const result = parsedJson.result;
-                        const results = result.results;
-                        if (results === undefined) {
-                            client.end();
-                            return;
-                        }
-                    } else {
-                        client.end();
-                        return;
-                    }
-                } catch (error) {
-                    console.error('Error parsing JSON:', error);
-                }
-            };
-        });
-        setTimeout(() => {
-            if (!client.destroyed) {
-                client.end();
-            }
-        }, 5000);*/
     });
     client.on('error', (error) => {
         console.error('Error sending profile data via Signal CLI:', error);
@@ -138,12 +123,35 @@ function setupbotprofile() {
 function main() {
     console.log('TritiumBot starting...');
     daemon = signalclidaemon();
-    const starthook = setInterval(() => {
-        if (fs.existsSync(socketpath)) {
-            clearInterval(starthook);
-            signalclihook();
-        }
-    }, 100);
+    if (socketpath.includes(':')) {
+        console.log('Method: TCP');
+        const starthook = setInterval(() => {
+            const testClient = new net.Socket();
+            testClient.setTimeout(1000);
+            
+            startconn(testClient, () => {
+                testClient.destroy();
+                clearInterval(starthook);
+                signalclihook();
+            });
+            
+            testClient.on('error', () => {
+                testClient.destroy();
+            });
+            
+            testClient.on('timeout', () => {
+                testClient.destroy();
+            });
+        }, 2000);
+    } else {
+        console.log('Method: UNIX Socket');
+        const starthook = setInterval(() => {
+            if (fs.existsSync(socketpath)) {
+                clearInterval(starthook);
+                signalclihook();
+            }
+        }, 100);
+    }
 };
 
 main();
